+++ test/network/network_connectivity_web_test.dart
@@ -1,4 +1,6 @@
 @TestOn('chrome')
+library;
+
 import 'dart:async';
 
 import 'package:test/test.dart';